+++ R/waiter.R
@@ -238,7 +238,7 @@ waiterPreloader <- function(
     "window.ran = false;",
     "$(document).on('shiny:idle', function(event){
       if(!window.ran)
-        waiter.hideWaiter(id = null);
+        waiter.hide(id = null);
 
       window.ran = true;
     });"
@@ -272,7 +272,7 @@ waiterHideOnRender <- function(id){
   script <- sprintf(
     "$(document).on('shiny:value', function(event) {
       if(event.name == '%s'){
-        waiter.hideWaiter(null);
+        waiter.hide(null);
       }
     });",
     id
@@ -316,7 +316,7 @@ waiterOnBusy <- function(html = spin_1(), color = "#333e48", logo = "", image = 
     });
     
     $(document).on('shiny:idle', function(event) {
-      waiter.hideWaiter(null);
+      waiter.hide(null);
     });"
   )
 
@@ -797,5 +797,60 @@ autoWaiter <- function(
   shiny::tagList(
     useWaiter(),
     HTML(sprintf("<script>%s</script>", script))
+  )
+}
+
+#' With Waiter
+#' 
+#' Adds a waiter to a rective UI element.
+#' Thew waiter is displayed when the element is
+#' invalidated then is removed when the element 
+#' receives a new value.
+#' 
+#' @inheritParams waiter
+#' @param element A reactive element, e.g.: `uiOutput`,
+#' or `plotOutput`.
+#' 
+#' @export 
+withWaiter <- function(
+  element,
+  html = spin_1(), 
+  color = "#333e48",
+  image = ""
+){
+  if(missing(element))
+    stop("Missing `element`", call. = FALSE)
+
+  id <- element$attribs$id
+  html <- as.character(html)
+  html <- gsub("\n", "", html)
+
+  script <- paste0(
+    "$(document).on('shiny:outputinvalidated', function(event) {
+      if(event.target.id != '", id, "')
+        return;
+
+      waiter.show({
+        id: '", id, "',
+        html: '", html, "', 
+        color: '", color, "',
+        image: '", image, "'
+      });
+    });
+    
+    $(document).on('shiny:value shiny:error', function(event) {
+      if(event.target.id != '", id, "')
+        return;
+      waiter.hide('", id, "');
+    });"
+  )
+
+  tagList(
+    singleton(
+      HTML(
+        paste0("<script>", script, "</script>")
+      )
+    ),
+    element
   )
 }